requirements.txt
numpy
scipy
pillow
opencv-python
torch
scikit-learn
easyocr
joblib
pandas
pyautogui